refactor(stdlib): split `GradientCast` and `DistanceTo`
